refactor: reorganize folder structure for some pages
