+++ src/pages/AdminLogsPage.jsx
@@ -10,7 +10,6 @@ function AdminLogsPage() {
     <div className="flex min-h-screen w-full">
       <BlogSidebarComponent />
       <div className="flex-1 p-4 overflow-x-auto">
-        <h1 className="text-3xl font-bold mb-6">Uygulama Loglama Yönetimi</h1>
         <LogViewer />
       </div>
     </div>
