+++ src/pages/auth/DeleteAccountPage.jsx
