--- src/pages/other_pages/PricePage.jsx
@@ -1,15 +1,0 @@
-import BannerComponent from "../../components/header/BannerComponent";
-import CustomNavbar from "../../components/header/CustomNavbar";
-import PriceTableComponent from "../../components/tanıtım_page/PriceTableComponent";
-
-function PricePage() {
-  return (
-    <>
-      <BannerComponent />
-      <CustomNavbar />
-      <PriceTableComponent />
-    </>
-  );
-}
-
-export default PricePage;
